3Dモデル変換アプリをリファクタリング
・一部システムをフォルダ内に移動して整理
    - サポート用のシステムのためのフォルダを設置、該当システムを移動
・一部の型の定義を各定義の専用のファイルに移動
    - 色データ用の型を移動
    - シェーダーデータ設定用クラスを移動
・頂点レイアウトの切り替え時に解放してから切り替えるように変更
