+++ プロジェクト/3Dモデル変換アプリ/Project/3D_Model_Converter_And_Drawer/3D_Model_Converter_And_Drawer.csproj
@@ -105,7 +105,8 @@
     <Compile Include="3D_Model_Importer\Import_System\CS_Animation_Model_Import_System.cs" />
     <Compile Include="3D_Model_Importer\Import_System\CS_Static_Model_Import_System.cs" />
     <Compile Include="Constant_Data\S_WVP.cs" />
-    <Compile Include="CS_File_Write_Load_System.cs" />
+    <Compile Include="Support_System\CS_Color.cs" />
+    <Compile Include="Support_System\CS_File_Write_Load_System.cs" />
     <Compile Include="d3d11\CS_DX_11_Constant_Buffer.cs" />
     <Compile Include="d3d11\CS_DX_11_Draw_Call_System.cs" />
     <Compile Include="d3d11\CS_DX_11_Texture.cs" />
@@ -156,12 +157,14 @@
     <Compile Include="Main_Form.Designer.cs">
       <DependentUpon>Main_Form.cs</DependentUpon>
     </Compile>
-    <Compile Include="CS_Main_Program.cs" />
-    <Compile Include="CS_My_Math_System.cs" />
+    <Compile Include="Support_System\CS_Main_Program.cs" />
+    <Compile Include="Support_System\CS_My_Math_System.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Rendering_Systems\Camera\CS_Camera_Manager.cs" />
     <Compile Include="Rendering_Systems\Camera\CS_Camera_System.cs" />
     <Compile Include="Rendering_Systems\Projection\CS_Projection.cs" />
+    <Compile Include="Support_System\C_Shader_Source.cs" />
+    <Compile Include="Support_System\E_BYTE_KIND.cs" />
     <Compile Include="UI\Announce_Bord\Form_Annouce_Bord.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -229,6 +232,7 @@
       <Install>false</Install>
     </BootstrapperPackage>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\AssimpNet.4.1.0\build\AssimpNet.targets" Condition="Exists('..\packages\AssimpNet.4.1.0\build\AssimpNet.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
