+++ プロジェクト/3Dモデル変換アプリ/Project/3D_Model_Converter_And_Drawer/Main_Form.cs
@@ -25,6 +25,7 @@ using _3D_Model_Converter_And_Drawer._3DModel;
 using _3D_Model_Converter_And_Drawer.d3d11.Transform;
 using _3D_Model_Converter_And_Drawer.Rendering_Systems.Camera;
 using _3D_Model_Converter_And_Drawer.d3d11;
+using _3D_Model_Converter_And_Drawer.Support_System;
 using SharpDX.Direct3D11;
 using SharpDX.D3DCompiler;
 
@@ -125,8 +126,16 @@ namespace _3D_Model_Converter_And_Drawer
 			uc_dx_11_panel.mp_shader.mp_shader_path = m_triangle_shader.mp_shader;
 
 
+			// 頂点レイアウトを初期化
+			if (uc_dx_11_panel.mp_shader.mp_layout != null)
+			{
+                uc_dx_11_panel.mp_shader.mp_layout.Dispose();
+                uc_dx_11_panel.mp_shader.mp_layout = null;
+            }
+
+
 			// 三角形の頂点レイアウトの生成
-			uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
+            uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
 			(
 				uc_dx_11_panel.mp_renderer.mp_device,
 				ShaderSignature.GetInputSignature(uc_dx_11_panel.mp_shader.mp_compiled_vertex_shader_code),
@@ -136,7 +145,7 @@ namespace _3D_Model_Converter_And_Drawer
 					new InputElement("COLOR", 0, SharpDX.DXGI.Format.R32G32B32A32_Float, 16, 0)
 				}
 			);
-
+			
 			return;
 		}
 
@@ -148,8 +157,16 @@ namespace _3D_Model_Converter_And_Drawer
 			uc_dx_11_panel.mp_shader.mp_shader_path = m_static_model_shader.mp_shader;
 
 
-			// 静的モデルの頂点レイアウトの生成
-			uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
+            // 頂点レイアウトを初期化
+            if (uc_dx_11_panel.mp_shader.mp_layout != null)
+            {
+                uc_dx_11_panel.mp_shader.mp_layout.Dispose();
+                uc_dx_11_panel.mp_shader.mp_layout = null;
+            }
+
+
+            // 静的モデルの頂点レイアウトの生成
+            uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
 			(
 				uc_dx_11_panel.mp_renderer.mp_device,
 				ShaderSignature.GetInputSignature(uc_dx_11_panel.mp_shader.mp_compiled_vertex_shader_code),
@@ -175,8 +192,16 @@ namespace _3D_Model_Converter_And_Drawer
 			uc_dx_11_panel.mp_shader.mp_shader_path = m_animation_model_shader.mp_shader;
 
 
-			// アニメーションモデルの頂点レイアウトの生成
-			uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
+            // 頂点レイアウトを初期化
+            if (uc_dx_11_panel.mp_shader.mp_layout != null)
+            {
+                uc_dx_11_panel.mp_shader.mp_layout.Dispose();
+                uc_dx_11_panel.mp_shader.mp_layout = null;
+            }
+
+
+            // アニメーションモデルの頂点レイアウトの生成
+            uc_dx_11_panel.mp_shader.mp_layout = new InputLayout
 			(
 				uc_dx_11_panel.mp_renderer.mp_device,
 				ShaderSignature.GetInputSignature(uc_dx_11_panel.mp_shader.mp_compiled_vertex_shader_code),
@@ -243,7 +268,6 @@ namespace _3D_Model_Converter_And_Drawer
 		// 指定されたパスのシェーダーをロードしセットする　引数：読み込んだシェーダー設定先のシェーダー設定用システム, シェーダーパス
 		private void M_Load_Shader_By_Path(out C_Shader_Source out_shader_source, string in_shader_path)
 		{
-			// シェーダーロード
 			out_shader_source = new C_Shader_Source();
 			out_shader_source.PropertyChanged += M_source_PropertyChanged;
 			out_shader_source.mp_shader = File.ReadAllText(Path.Combine(Path.GetDirectoryName(Application.ExecutablePath), in_shader_path), Encoding.UTF8);
@@ -309,22 +333,22 @@ namespace _3D_Model_Converter_And_Drawer
 			switch (byte_shrink_times)
 			{
 				// バイト表記
-				case 0:
+				case (int)E_BYTE_KIND.e_BYTE:
 					result_string += con_BYTE;
 					break;
 
 				// キロバイト表記
-				case 1:
+				case (int)E_BYTE_KIND.e_KILO_BYTE:
 					result_string += con_KILO_BYTE;
 					break;
 
 				// メガバイト表記
-				case 2:
+				case (int)E_BYTE_KIND.e_MEGA_BYTE:
 					result_string += con_MEGA_BYTE;
 					break;
 
 				// ギガバイト表記
-				case 3:
+				case (int)E_BYTE_KIND.e_GIGA_BYTE:
 					result_string += con_GIGA_BYTE;
 					break;
 			}
@@ -715,15 +739,15 @@ namespace _3D_Model_Converter_And_Drawer
 			if (m_animation_model == null)
 			{
 				return;
-			}
+            }
 
 
-			// 描画用の設定を初期化する
-			m_draw_setting = new CS_DX_11_Draw_Call_System();
+            // ☆ 変数宣言 ☆ //
+            CS_DX_11_Vertex_Buffer<S_Animation_Vertex_Data> new_vertex_buffer = new CS_DX_11_Vertex_Buffer<S_Animation_Vertex_Data>(); // 頂点バッファ
 
 
-			// ☆ 変数宣言 ☆ //
-			CS_DX_11_Vertex_Buffer<S_Animation_Vertex_Data> new_vertex_buffer = new CS_DX_11_Vertex_Buffer<S_Animation_Vertex_Data>(); // 頂点バッファ
+            // 描画用の設定を初期化する
+            m_draw_setting = new CS_DX_11_Draw_Call_System();
 
 
 			// アニメーションモデル用のシェーダーをセット
@@ -992,86 +1016,5 @@ namespace _3D_Model_Converter_And_Drawer
 
 			return;
 		}
-	}
-
-
-	// シェーダー設定用のクラス
-	class C_Shader_Source : INotifyPropertyChanged
-	{
-		// ☆ 変数宣言 ☆ //
-		string m_shader_path;   // シェーダーファイルのパス
-
-
-		// ☆ プロパティ ☆ //
-
-		// シェーダーファイルのパス
-		public string mp_shader
-		{
-			// ゲッタ
-			get
-			{
-				return m_shader_path;
-			}
-
-			// セッタ
-			set
-			{
-				if (m_shader_path == value)
-				{
-					return;
-				}
-
-				m_shader_path = value;
-				RaisePropertyChanged(nameof(mp_shader));
-
-				return;
-			}
-		}
-
-
-		// ☆ イベント ☆ //
-		public event PropertyChangedEventHandler PropertyChanged;   // プロパティ変更時
-
-
-		// プロパティ変更時
-		void RaisePropertyChanged(string prop)
-		{
-			// ☆ 変数宣言 ☆ //
-			var handler = PropertyChanged;  // プロパティ変更イベントのハンドル
-
-
-			// プロパティ変更イベントが取得できたならプロパティ変更イベントを実行
-			if (handler != null)
-			{
-				handler(this, new PropertyChangedEventArgs(prop));
-			}
-		}
-	}
-
-
-	// ロード情報表示用のクラス
-	static class CS_Color_Extensions
-	{
-		// SharpDXのカラーをfloat4のカラーに変換
-		public static System.Drawing.Color M_SharpDX_Color_To_Float4_Color(this Color4 src)
-		{
-			return System.Drawing.Color.FromArgb(
-				(byte)(src.Alpha * 255),
-				(byte)(src.Red * 255),
-				(byte)(src.Green * 255),
-				(byte)(src.Blue * 255)
-				);
-		}
-
-		// float4のカラーをSharpDXのカラーに変換
-		public static Color4 M_Float4_Color_To_SharpDX_Color(this System.Drawing.Color value)
-		{
-			return new Color4(
-				value.R / 255.0f,
-				value.G / 255.0f,
-				value.B / 255.0f,
-				value.A / 255.0f
-				);
-		}
-	}
+    }
 }
