+++ プロジェクト/3Dモデル変換アプリ/Project/3D_Model_Converter_And_Drawer/Support_System/E_BYTE_KIND.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace _3D_Model_Converter_And_Drawer.Support_System
+{
+    // ☆ 列挙 ☆ //
+
+    // バイトの種類の列挙
+    enum E_BYTE_KIND
+    {
+        e_BYTE, // バイト
+        e_KILO_BYTE, // キロバイト
+        e_MEGA_BYTE, // メガバイト
+        e_GIGA_BYTE, // ギガバイト
+        e_TERA_BYTE, // テラバイト
+    }
+}
